refactor: 스타일, 타입, 컴포넌트 분리
