+++ client/src/hooks/useModal.style.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+const ModalContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 11;
+  margin: 0px;
+`;
+
+const ModalBackground = styled.div`
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.75);
+`;
+
+export { ModalContainer, ModalBackground };
