+++ client/src/hooks/useModal.tsx
@@ -1,27 +1,7 @@
 import React, { useState } from "react";
-import styled from "styled-components";
 import { createPortal } from "react-dom";
-
-const ModalContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 11;
-  margin: 0px;
-`;
-
-const ModalBackground = styled.div`
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.75);
-`;
-
-interface ModalPortalProps {
-  children: React.ReactNode;
-}
+import { ModalContainer, ModalBackground } from "./useModal.style";
+import ModalPortalProps from "./useModal.type";
 
 const useModal = () => {
   const [modalOpen, setModalOpen] = useState(false);
