+++ client/src/hooks/useModal.type.ts
@@ -1,0 +1,5 @@
+interface ModalPortalProps {
+  children: React.ReactNode;
+}
+
+export default ModalPortalProps;
