+++ client/src/hooks/useOutsideDetecter.type.tsx
@@ -1,0 +1,6 @@
+interface OutsideDetecterProps {
+  children: React.ReactNode;
+  callback: () => void;
+}
+
+export default OutsideDetecterProps;
