+++ client/src/pages/Manage/DeleteSurveyModal.component.tsx
@@ -1,34 +1,12 @@
 import React from "react";
-import styled from "styled-components";
 import axios from "axios";
+import { Container } from "./DeleteSurveyModal.style";
+import DeleteSurveyModalProps from "./DeleteSurveyModal.type";
 
-const Container = styled.div`
-  position: absolute;
-  top: 35%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-
-  width: 400px;
-  border-radius: 9px;
-  padding: 20px;
-
-  z-index: 2;
-  background-color: white;
-`;
-
-function DeleteSurveyModal({
-  closeModal,
-  modifyListBySurveyDelete,
-  selectedSurvey,
-}: {
-  closeModal: () => void;
-  selectedSurvey: { id: string; index: number };
-  modifyListBySurveyDelete: (index: number) => void;
-}) {
-  console.log(selectedSurvey);
+function DeleteSurveyModal({ closeModal, renderByDeleteForm, selectedSurvey }: DeleteSurveyModalProps) {
   const onClickDeleteSurvey = async () => {
     await axios.delete(`http://localhost:8080/api/forms/${selectedSurvey.id}`);
-    modifyListBySurveyDelete(selectedSurvey.index);
+    renderByDeleteForm(selectedSurvey.index);
     closeModal();
   };
