+++ client/src/pages/Manage/DeleteSurveyModal.type.ts
@@ -1,0 +1,7 @@
+interface DeleteSurveyModalProps {
+  closeModal: () => void;
+  selectedSurvey: { id: string; index: number };
+  renderByDeleteForm: (index: number) => void;
+}
+
+export default DeleteSurveyModalProps;
