+++ client/src/pages/Manage/EditNameModal.type.ts
@@ -1,0 +1,7 @@
+interface EditNameModalProps {
+  closeModal: () => void;
+  selectedSurvey: { id: string; index: number };
+  renderByNameChange: (index: number, title: string) => void;
+}
+
+export default EditNameModalProps;
